requirements.txt
aiogram>=3.16.0
python-dotenv>=1.0.1
selenium>=4.27.1